refactor(product): domain-service에 있으면 안되는 로직 usecase로 이동
