+++ src/product/application/use-cases/tier-1-in-domain/release-stock.use-case.spec.ts
@@ -7,7 +7,7 @@ import {
 import { StockReservation } from "@/product/domain/entities/stock-reservation.entity";
 import { Product } from "@/product/domain/entities/product.entity";
 import { v4 as uuidv4 } from "uuid";
-import { ReleaseStockDomainService } from "@/product/domain/services/release-stock.service";
+import { ValidateStockService } from "@/product/domain/services/validate-stock.service";
 
 jest.mock("typeorm-transactional", () => ({
   Transactional: () => () => ({}),
@@ -32,7 +32,7 @@ describe("ReleaseStockUseCase", () => {
     const module = await Test.createTestingModule({
       providers: [
         ReleaseStockUseCase,
-        ReleaseStockDomainService,
+        ValidateStockService,
         {
           provide: "ProductRepositoryInterface",
           useValue: productRepository,
