+++ src/product/application/use-cases/tier-1-in-domain/reserve-stock.use-case.spec.ts
@@ -9,7 +9,7 @@ import {
 import { StockReservation } from "@/product/domain/entities/stock-reservation.entity";
 import { Product } from "@/product/domain/entities/product.entity";
 import { v4 as uuidv4 } from "uuid";
-import { ReserveStockDomainService } from "@/product/domain/services/reserve-stock.service";
+import { ValidateStockService } from "@/product/domain/services/validate-stock.service";
 
 jest.mock("typeorm-transactional", () => ({
   Transactional: () => () => ({}),
@@ -33,7 +33,7 @@ describe("ReserveStockUseCase", () => {
     const module = await Test.createTestingModule({
       providers: [
         ReserveStockUseCase,
-        ReserveStockDomainService,
+        ValidateStockService,
         {
           provide: "ProductRepositoryInterface",
           useValue: productRepository,
