--- src/product/domain/services/release-stock.service.ts
@@ -1,22 +1,0 @@
-import { Product } from "../entities/product.entity";
-import { StockReservation } from "../entities/stock-reservation.entity";
-import { StockReservationNotActiveError } from "../exceptions/product.exceptions";
-
-export class ReleaseStockDomainService {
-  async releaseStock({
-    stockReservation,
-    product,
-    idempotencyKey,
-  }: {
-    stockReservation: StockReservation;
-    product: Product;
-    idempotencyKey: string;
-  }): Promise<void> {
-    if (!stockReservation.isActive) {
-      throw new StockReservationNotActiveError(stockReservation.id);
-    }
-
-    product.releaseStock(stockReservation.quantity);
-    stockReservation.releaseStock(idempotencyKey);
-  }
-}
